+++ pentagon/component/__init__.py
@@ -1,5 +1,4 @@
 import os
-import jinja2
 import glob
 import shutil
 import logging
@@ -15,11 +14,11 @@ class ComponentBase(object):
     """ Base class for Pentagon Components. """
     _required_parameters = []
 
-# List of environment variables to use.
-# If set, they should override other data sources.
-# Lower Case here will find upper case environment variables.
-# If a dictionary is passed, the key is the variable name used in context,
-# and the value is the environment variable name.
+    # List of environment variables to use.
+    # If set, they should override other data sources.
+    # Lower Case here will find upper case environment variables.
+    # If a dictionary is passed, the key is the variable name used in context,
+    # and the value is the environment variable name.
     _environment = []
     _defaults = {}
 
@@ -36,7 +35,9 @@ class ComponentBase(object):
                 missing_parameters.append(item)
 
         if missing_parameters:
-            logging.error("Missing required data parameters: {}".format(",".join(missing_parameters)))
+            logging.error("Missing required data parameters: {}".format(
+                ", ".join(missing_parameters)))
+            logging.error("You can set parameters with '-Dparam_name=value'.")
             sys.exit(1)
 
     @property
@@ -73,17 +74,20 @@ class ComponentBase(object):
         try:
             class_name = self.__class__.__name__.lower()
             pentagon_defaults = getattr(PentagonDefaults, class_name)
-            logging.debug("Adding Pentagon Defaults Last {}".format(pentagon_defaults))
+            logging.debug(
+                "Adding Pentagon Defaults Last {}".format(pentagon_defaults))
             self._merge_data(pentagon_defaults)
         except AttributeError, e:
-            logging.info("No top level defaults for Pentagon component {} ".format(class_name.lower()))
+            logging.info("No top level defaults for Pentagon component {} ".format(
+                class_name.lower()))
 
     def _render_directory_templates(self):
         """ Loop and use render_template helper method on all templates in destination directory  """
         template_location = self._destination_directory_name
-        if os.path.isfile(self._destination_directory_name):
-            template_location = os.path.dirname(self._destination_directory_name)
-            logging.debug("{} is a file. Using the directory {} instead.".format(self._destination_directory_name, template_location))
+        if os.path.isfile(template_location):
+            template_location = os.path.dirname(template_location)
+            logging.debug("{} is a file. Using the directory {} instead.".format(
+                self._destination_directory_name, template_location))
         logging.debug("Rendering Templates in {}".format(template_location))
         for folder, dirnames, files in os.walk(template_location):
             for template in glob.glob(folder + "/*.jinja"):
@@ -92,37 +96,70 @@ class ComponentBase(object):
                 path = '/'.join(template.split('/')[0:-1])
                 target_file_name = re.sub(r'\.jinja$', '', template_file_name)
                 target = folder + "/" + target_file_name
-                render_template(template_file_name, path, target, self._data, overwrite=self._overwrite)
+                render_template(template_file_name, path, target,
+                                self._data, overwrite=self._overwrite)
 
     def _remove_init_file(self):
         """ delete init file, if it exists from template target directory """
 
         for root, dirs, files in os.walk(self._destination_directory_name):
             for name in files:
-                if "__init__.py" == name:
-                    logging.debug('Removing: {}'.format(os.path.join(root, name)))
+                if "__init__.py" == name or "__init__.pyc" == name:
+                    logging.debug('Removing: {}'.format(
+                        os.path.join(root, name)))
                     os.remove(os.path.join(root, name))
 
     def _merge_data(self, new_data, clobber=False):
-        """ accepts new_data (dict) and clobbber (boolean). Merges dictionary with existing instance dictionary _data. If clobber is True, overwrites value. Defaults to false """
+        """ accepts new_data (dict) and clobber (boolean). Merges dictionary with existing instance dictionary _data. If clobber is True, overwrites value. Defaults to false """
         for key, value in new_data.items():
             if self._data.get(key) is None or clobber:
-                logging.debug("Setting component data {}: {}".format(key, value))
+                logging.debug(
+                    "Setting component data {}: {}".format(key, value))
                 self._data[key] = value
 
     def add(self, destination, overwrite=False):
         self._destination = destination
         self._overwrite = overwrite
+        self._display_settings_to_user()
         try:
+            # Add all files from the component templates to the destination directory
             self._add_files()
+            # Remove any __init__ files in the destination that were copied from the component templates
             self._remove_init_file()
+            # For all the jinja templates in the destination directory, render them
             self._render_directory_templates()
-            logging.info("New component added. Source your environment before proceeding or unexpected behavior may result.")
+            logging.info("New component added. Source your environment before "
+                         "proceeding or unexpected behavior may result.")
         except Exception as e:
-            logging.error("Error occured configuring component")
+            logging.error("Error occurred configuring component")
             logging.error(e)
             logging.debug(traceback.format_exc(e))
             sys.exit(1)
+
+    def _display_settings_to_user(self):
+        logging.info("Pentagon will write to the following directory: "
+                     "(set with '-o ./path')")
+        logging.info("  Path: \"{}\"".format(self._destination_directory_name))
+        logging.info("Displaying provided and default values for this component: "
+                     "(e.g. '-Dparam_name=abcd')")
+        for key, value in self._data.iteritems():
+            using_defaults = False
+            if key in self._defaults.keys():
+                if self._data[key] == self._defaults[key]:
+                    using_defaults = True
+
+            is_default = "(Default Value)" if using_defaults else ""
+            logging.info("  {0:34} = {1:20} {2}".format(
+                key,
+                value,
+                is_default,
+            ))
+        response = raw_input("This look ok to proceed? (y/N) ")
+        if response == 'y' or response == 'Y':
+            return
+        else:
+            logging.info("Exiting because you did not accept the inputs.")
+            exit()
 
     def _add_files(self, sub_path=None):
         """ Copies files and templates from <component>/files """
@@ -135,7 +172,8 @@ class ComponentBase(object):
         else:
             source = self._files_directory
 
-        logging.debug("Adding file: {} -> {}".format(source, self._destination_directory_name))
+        logging.debug("Adding file: {} -> {}".format(source,
+                                                     self._destination_directory_name))
         if os.path.isfile(source):
             shutil.copy(source, self._destination_directory_name)
         elif os.path.isdir(source):