+++ pentagon/component/gcp/network.py
@@ -1,7 +1,0 @@
-from pentagon.component import ComponentBase
-import os
-
-
-class Network(ComponentBase):
-
-    _path = os.path.dirname(__file__)