+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.0.0
+### Added
+- Support for GCP / GKE terraform templates on inventory init
+
+### Changed
+- Now all pentagon runs will confirm all the values that are set and what the values are set to (one step closer to better transparency)
+
 ## 2.7.3
 ### Fixed
 - missing imports for latest migration